Refactor endpoints to introduce versioning
Renamed CreateBrand and GetBrandById endpoints to include versioning. Updated the API program configuration to support versioning with a default version of 1 and added Swagger documentation for version 1.
